complete json processing feature

+++ payslip_functions.py
@@ -18,10 +18,10 @@ def number_cruncher(personal_info_set, calculated_tax):
     f_name, l_name, annual_salary, super_rate, pay_start_date = personal_info_set
     name = f_name + ' ' + l_name
     pay_period = parse_date(pay_start_date)
-    gross_income = round(float(annual_salary) / 12)
-    income_tax = round(calculated_tax / 12)
+    gross_income = int(round(float(annual_salary) / 12))
+    income_tax = int(round(calculated_tax / 12))
     net_income = gross_income - income_tax
-    superannuation = round(float(gross_income) * float(super_rate) / 100)
+    superannuation = int(round(float(gross_income) * float(super_rate) / 100))
 
     return name, pay_period, gross_income, income_tax, net_income, superannuation
